SNC-2: Refactored code; added viewmodel; added hilt.
