+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <application
+        android:name=".App"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -12,7 +13,7 @@
         android:theme="@style/Theme.SocialNetworkClient"
         tools:targetApi="33">
         <activity
-            android:name=".presentation.ui.AuthActivity"
+            android:name=".presentation.ui.signup.AuthActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.SocialNetworkClient">
