+++ app/src/main/java/com/stanislavdumchykov/socialnetworkclient/App.kt
@@ -1,0 +1,7 @@
+package com.stanislavdumchykov.socialnetworkclient
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class App : Application()
