+++ app/src/main/java/com/stanislavdumchykov/socialnetworkclient/presentation/navigation/NavGraph.kt
@@ -4,9 +4,9 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.stanislavdumchykov.socialnetworkclient.presentation.ui.ContactList
-import com.stanislavdumchykov.socialnetworkclient.presentation.ui.MyProfile
-import com.stanislavdumchykov.socialnetworkclient.presentation.ui.SignUpScreen
+import com.stanislavdumchykov.socialnetworkclient.presentation.ui.contactlist.ContactList
+import com.stanislavdumchykov.socialnetworkclient.presentation.ui.myprofile.MyProfile
+import com.stanislavdumchykov.socialnetworkclient.presentation.ui.signup.SignUpScreen
 
 @Composable
 fun SetupNavGraph(navController: NavHostController) {
@@ -26,7 +26,7 @@ fun SetupNavGraph(navController: NavHostController) {
         }
         composable(
             route = Routes.ContactList.route
-        ){
+        ) {
             ContactList(navController = navController)
         }
     }
