+++ build.gradle
@@ -1,12 +1,18 @@
 buildscript {
     ext {
         compose_ui_version = '1.4.1'
+        hilt_version = '2.45'
+    }
+
+    dependencies {
+        classpath "com.google.dagger:hilt-android-gradle-plugin:$hilt_version"
     }
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
     id 'com.android.application' version '7.4.2' apply false
     id 'com.android.library' version '7.4.2' apply false
     id 'org.jetbrains.kotlin.android' version '1.8.10' apply false
+    id 'com.google.dagger.hilt.android' version '2.45' apply false
 }
 
 task clean(type: Delete) {
